\name{animint-package}
\alias{animint-package}
\docType{package}
\title{Interactive animations}
\description{Interactive animations can be defined using the
clickSelects and showSelected aesthetics.}
\details{
\tabular{ll}{Package: \tab animint\cr
Maintainer: \tab Toby Dylan Hocking <toby@sg.cs.titech.ac.jp>\cr
Author: \tab Toby Dylan Hocking\cr
Version: \tab 1.0\cr
License: \tab GPL-3\cr
Title: \tab Interactive animations\cr
Depends: \tab RJSONIO, ggplot2, R (>= 2.10), proto, grid\cr}
}
\author{Toby Dylan Hocking}

\keyword{ package }

