% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animint.R
\name{animint}
\alias{animint}
\title{animint}
\usage{
animint(...)
}
\arguments{
\item{...}{passed to list}
}
\value{
a list of class "animint"
}
\description{
Create a new animint, which is a list of ggplots and options that
has class "animint"
}
\author{
Toby Dylan Hocking
}

