% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gist.R
\name{animint2gist}
\alias{animint2gist}
\title{Convert a list of ggplots to an interactive animation and post files as a gist}
\usage{
animint2gist(plot.list, out.dir = tempfile(), json.file = "plot.json",
  css.file = "", open.browser = interactive(),
  url_prefix = "http://bl.ocks.org", description = plot.list$title,
  screenshot = TRUE, ...)
}
\arguments{
  \item{plot.list}{a named list of ggplots and option
  lists.}

  \item{out.dir}{local directory to store html/js/csv
  files.}

  \item{json.file}{character string that names the JSON
  file with metadata associated with the plot.}

  \item{url_prefix}{first part of URL for viewing result.}

  \item{description}{Brief description of gist. This
  becomes the plot title on the bl.ocks/username page.}

  \item{screenshot}{logical. If TRUE, attempt to take
  screenshot of result and include with gist.}

  \item{...}{options passed onto \link{gistr::gist_create}}
}
\description{
Before using this function set your appropriate
'github.username' and 'github.password' \link{options}
}
\examples{
\dontrun{
library(animint)
iris$id <- 1:nrow(iris)
viz <- list(petal=ggplot()+
         geom_point(aes(Petal.Width, Petal.Length, fill=Species,
                        clickSelects=id), data=iris),
       sepal=ggplot()+
         geom_point(aes(Sepal.Width, Sepal.Length, fill=Species,
                        clickSelects=id), data=iris))
animint2gist(viz, description = "My animint plot")
}
}

