% Generated by roxygen2 (4.0.0): do not edit by hand
\name{animint2HTML}
\alias{animint2HTML}
\title{Run gg2animint and then applyJS.}
\usage{
animint2HTML(gg.opt.list)
}
\arguments{
  \item{gg.opt.list}{List of ggplots and option lists.}
}
\value{
HTML generated after evaluation of JavaScript.
}
\description{
Run gg2animint and then applyJS.
}
\author{
Toby Dylan Hocking
}

