% Generated by roxygen2 (4.0.0): do not edit by hand
\name{animint2serv}
\alias{animint2serv}
\title{Run gg2animint in temporary directory and serve files on localhost}
\usage{
animint2serv(gg.opt.list, port = 4321)
}
\arguments{
  \item{gg.opt.list}{List of ggplots and option lists.}

  \item{port}{the TCP port number; by default it is 4321.}
}
\value{
HTML generated after evaluation of JavaScript.
}
\description{
This function uses a seperate R session to serve the
relevant files (via the servr package).
}
\author{
Carson Sievert
}

