% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.R
\name{animintOutput}
\alias{animintOutput}
\title{Shiny ui output function}
\usage{
animintOutput(outputId)
}
\arguments{
\item{outputId}{output variable to read the plot from}
}
\description{
Shiny ui output function
}
\seealso{
http://shiny.rstudio.com/articles/building-outputs.html
}

