% Generated by roxygen2 (4.0.0): do not edit by hand
\name{applyJS}
\alias{applyJS}
\title{Evaluate JavaScript code for testing purposes}
\usage{
applyJS(html.file)
}
\arguments{
  \item{html.file}{HTML file containing <script> tags.}
}
\value{
HTML+SVG result after applying the JavaScript code.
}
\description{
Evaluate JavaScript code for testing purposes
}
\author{
Toby Dylan Hocking
}

