\name{ChromHMMiterations}
\alias{ChromHMMiterations}
\docType{data}
\title{
  ChromHMM iterations
}
\description{
  ChromHMM was run for one iPS sample for 100 iterations
  and its convergence metrics and parameter values were saved
  at each iteration.
}
\usage{data("ChromHMMiterations")}
\format{
  Named list of 3 data.frames: metrics, transition, emission.
  }
\source{
  https://github.com/tdhock/ChromHMM-viz/blob/master/iterations.R
}
\keyword{datasets}
