\name{FunctionalPruning}
\alias{FunctionalPruning}
\docType{data}
\title{
  Functional Pruning Algorithm
}
\description{
  Output of an algorithm for optimal change-point detection.
}
\usage{data("FunctionalPruning")}
\format{
  a named list of 4 data.frames
}
\source{
https://github.com/tdhock/PeakSegFPOP-paper/blob/master/figure-constrained-PDPA-normal-real.R
}
\keyword{datasets}
