\name{geom_tallrect}
\alias{geom_tallrect}
\title{geom tallrect}
\description{ggplot2 geom with xmin and xmax aesthetics that covers the entire
y range, useful for clickSelects background elements.}
\usage{geom_tallrect(mapping = NULL, data = NULL, stat = "identity", 
    position = "identity", ...)}
\arguments{
  \item{mapping}{
}
  \item{data}{
}
  \item{stat}{
}
  \item{position}{
}
  \item{\dots}{
}
}



\author{Toby Dylan Hocking}





