% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_widerect}
\alias{geom_widerect}
\title{ggplot2 geom with ymin and ymax aesthetics that covers the entire x range, useful for clickSelects background elements.}
\usage{
geom_widerect(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{data set}

\item{stat}{statistic mapping, defaults to identity}

\item{position}{position mapping, defaults to identity}

\item{...}{other arguments}
}
\value{
ggplot2 layer
}
\description{
ggplot2 geom with ymin and ymax aesthetics that covers the entire x range, useful for clickSelects background elements.
}
\examples{
 \dontrun{ 
   source(system.file("examples/WorldBank.R", package = "animint"))
 }
}

