% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.R
\name{gg2animint_knitr}
\alias{gg2animint_knitr}
\title{Insert an interactive animation into an R markdown document.}
\usage{
gg2animint_knitr(plot.list)
}
\arguments{
  \item{plot.list}{named list of ggplots and option lists to
  pass to gg2animint.}
}
\value{
Nothing, but outputs HTML that will render the interactive
animation.
}
\description{
Insert an interactive animation into an R markdown document.
}
\author{
Toby Dylan Hocking
}

