% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/animint.R
\name{is.rgb}
\alias{is.rgb}
\title{Check if character is an RGB hexadecimal color value}
\usage{
is.rgb(x)
}
\arguments{
  \item{x}{character}
}
\value{
True/False value
}
\description{
Check if character is an RGB hexadecimal color value
}

