% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/animint.R
\name{is.showSelected}
\alias{is.showSelected}
\title{Test if aesthetics are showSelected.}
\usage{
is.showSelected(x)
}
\arguments{
\item{x}{character vector.}
}
\value{
logical vector
}
\description{
Test if aesthetics are showSelected.
}
\author{
Toby Dylan Hocking
}

