% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knitr.R
\name{knit_print.animint}
\alias{knit_print.animint}
\title{Insert an interactive animation into an R markdown document using a customized print method.}
\usage{
\method{knit_print}{animint}(x, options, ...)
}
\arguments{
  \item{x}{named list of ggplots and option lists to pass
  to gg2animint.}
}
\description{
Insert an interactive animation into an R markdown document
using a customized print method.
}
\author{
Carson Sievert
}
\references{
https://github.com/yihui/knitr/blob/master/vignettes/knit_print.Rmd
}

