% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoms.R
\name{make_tallrect_or_widerect}
\alias{make_tallrect_or_widerect}
\title{Make a clickSelects geom_widerect or geom_tallrect that completely
tiles the x or y range. This function is used internally by
make_tallrect or make_widerect, which are more user-friendly.}
\usage{
make_tallrect_or_widerect(aes.prefix, geom_xrect, data, var.name,
  even = FALSE, alpha = 0.5, ..., data.fun = identity)
}
\arguments{
\item{aes.prefix}{"x" or "y"}

\item{geom_xrect}{geom_tallrect or geom_widerect}

\item{data}{data.frame to analyze for unique var.name values.}

\item{var.name}{variable to be used for clickSelects}

\item{even}{Logical parameter, should xrects be of even width?}

\item{alpha}{transparency of a selected xrect, default 1/2.}

\item{...}{passed to geom_xrect}

\item{data.fun}{called on data passed to geom_xrect(aes(..),
data.fun(df)) this is useful in facetted plots, for adding
columns to the data.frame, if you want that geom in only one
panel.}
}
\value{
a geom_xrect layer
}
\description{
Make a clickSelects geom_widerect or geom_tallrect that completely
tiles the x or y range. This function is used internally by
make_tallrect or make_widerect, which are more user-friendly.
}
\author{
Toby Dylan Hocking
}

