% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geoms.R
\name{make_widerect}
\alias{make_widerect}
\title{Make a clickSelects geom_widerect that completely tiles the y
range. This makes it easy to construct widerects for the common
case of selecting a particular y value.}
\usage{
make_widerect(data, y.name, even = FALSE, alpha = 0.5, ...)
}
\arguments{
\item{data}{data.frame to analyze for unique y.name values.}

\item{y.name}{variable to be used for y, clickSelects.}

\item{even}{Logical parameter, should widerects be of even width?}

\item{alpha}{transparency of a selected widerect, default 1/2.}

\item{...}{passed to geom_widerect.}
}
\value{
a geom_widerect layer.
}
\description{
Make a clickSelects geom_widerect that completely tiles the y
range. This makes it easy to construct widerects for the common
case of selecting a particular y value.
}
\author{
Toby Dylan Hocking
}

