% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{makeDocs}
\alias{makeDocs}
\title{Make documentation web site.}
\usage{
makeDocs(doc.dir)
}
\arguments{
\item{doc.dir}{containing example subdirectories.}
}
\value{
nothing.
}
\description{
Make documentation web site.
}
\author{
Toby Dylan Hocking
}

