% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animint.R
\name{merge_recurse}
\alias{merge_recurse}
\title{Merge a list of data frames.}
\usage{
merge_recurse(dfs)
}
\arguments{
\item{dfs}{list of data frames}
}
\value{
data frame
}
\description{
Merge a list of data frames.
}

