\name{mixtureKNN}
\alias{mixtureKNN}
\docType{data}
\title{
  K-Nearest-Neighbors model of the mixture example data
}
\description{
  Several KNN models were fit to the mixture example data set.
}
\usage{data("mixtureKNN")}
\format{
  Named list of 9 data.frames.
}
\source{
  animint/inst/examples/knn.R
}
\references{
  help(mixture.example, package="ElemStatLearn")
}
\keyword{datasets}
