\name{pirates}
\alias{pirates}
\docType{data}
\title{
  Worldwide pirate attacks from 1978 to 2013
}
\description{
  Each row is 1 pirate attack, with geographic location and
  textual descriptions.
}
\usage{data(pirates)}
\format{
  A data frame with 6636 observations on the following 14 variables.
  \describe{
    \item{\code{Reference}}{factor}
    \item{\code{DateOfOcc}}{a Date}
    \item{\code{SubReg}}{factor}
    \item{\code{Aggressor}}{factor}
    \item{\code{Victim}}{factor}
    \item{\code{Descript}}{factor}
    \item{\code{Desc1}}{factor}
    \item{\code{Desc2}}{factor}
    \item{\code{Desc3}}{factor}
    \item{\code{Desc4}}{factor}
    \item{\code{Desc5}}{factor}
    \item{\code{Desc6}}{factor}
    \item{\code{coords.x1}}{a numeric vector}
    \item{\code{coords.x2}}{a numeric vector}
  }
}
\source{
http://msi.nga.mil/MSISiteContent/StaticFiles/Files/ASAM_shp.zip
}
\keyword{datasets}
