\name{PredictedPeaks}
\alias{PredictedPeaks}
\docType{data}
\title{
 Predictions from PeakSegJoint
}
\description{
  Chromosomes, peak locations across 393 samples.
}
\usage{data("PredictedPeaks")}
\format{List of data.frames.}
\source{
  https://github.com/tdhock/PeakSegJoint/blob/master/exec/Step6-write-bed-files.R
}
\keyword{datasets}
