\name{prior}
\alias{prior}
\docType{data}
\title{
Class prior change
}
\description{
Performance of several class prior change models was studied across 4
data sets.
}
\usage{data(prior)}
\format{
List of 2 data.frames: accuracy contains the mean and standard error of
the performance measures (sqErr and accuracy), data.set.info contains
meta-data about the dimension and number of positive and negative
examples in each data set.
}
\source{
MC du Plessis, personal communication.
}
\references{
  du Plessis MC, Sugiyama M.
  Semi-Supervised Learning of Class Balance under Class-Prior Change by
  Distribution Matching. ICML 2012.
}
\keyword{datasets}
