% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knitr.R
\name{renderAnimint}
\alias{renderAnimint}
\title{Create an animint output element}
\usage{
renderAnimint(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
  \item{expr}{}

  \item{env}{}

  \item{quoted}{}
}
\description{
Shiny server output function customized for animint plots
(similar to \code{shiny::plotOutput} and friends).
}
\seealso{
http://shiny.rstudio.com/articles/building-outputs.html
}

