% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{run_tests}
\alias{run_tests}
\title{Run animint tests}
\usage{
run_tests(browserName = "phantomjs", dir = ".", ..., filter = NULL,
  show.ps = FALSE)
}
\arguments{
  \item{browserName}{Name of the browser to use for
  testing. See ?RSelenium::remoteDriver for details.}

  \item{dir}{character string with the path to animint's
  source code.}

  \item{...}{passed onto RSelenium::remoteDriver}

  \item{filter}{If not NULL, only tests with file names
  matching this regular expression will be executed.
  Matching will take on the file name after it has been
  stripped of "test-" and ".r".}

  \item{show.ps}{show user invoked processes}

  \item{port}{port number used for local file server}
}
\description{
Convenience function for running animint tests.
}
\examples{
# testing requires that you load testthat first.
library("testthat")
# run all tests with phantomjs
run_tests()
# run tests in test-rotate.R with Firefox
run_tests("firefox", filter = "rotate")
run_tests("chrome")
}

