% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/animint.R
\name{saveLayer}
\alias{saveLayer}
\title{Save a layer to disk, save and return meta-data.}
\usage{
saveLayer(l, d, meta)
}
\arguments{
  \item{l}{one layer of the ggplot object.}

  \item{d}{one layer of calculated data from
  ggplot2::ggplot_build(p).}

  \item{meta}{environment of meta-data.}
}
\value{
list representing a layer, with corresponding aesthetics,
ranges, and groups.
}
\description{
Save a layer to disk, save and return meta-data.
}

