% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/animint.R
\name{selector.aes}
\alias{selector.aes}
\title{Parse selectors from aes names.}
\usage{
selector.aes(a.list)
}
\arguments{
\item{a.vec}{character vector of aes names.}
}
\value{
list of selector info.
}
\description{
Parse selectors from aes names.
}
\author{
Toby Dylan Hocking
}

