% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animint.R
\name{split.x}
\alias{split.x}
\title{Split data.frame into recursive list of data.frame.}
\usage{
\method{split}{x}(x, vars)
}
\arguments{
\item{x}{data.frame.}

\item{vars}{character vector of variable names to split on.}
}
\value{
recursive list of data.frame.
}
\description{
Split data.frame into recursive list of data.frame.
}

