% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{tests_exit}
\alias{tests_exit}
\title{Shut down external processes necessary for running tests}
\usage{
tests_exit(envir)
}
\arguments{
  \item{envir}{an environmnent.}
}
\value{
invisible(FALSE) if an error occurred; invisible(TRUE)
otherwise
}
\description{
Shuts down the local file server and remote driver
initiated by \link{tests_init()}.
}
\seealso{
\link{tests_run()}
}

