% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{tests_init}
\alias{tests_init}
\title{Initiate external processes necessary for running tests.}
\usage{
tests_init(browserName = "phantomjs", dir = ".", port = 4848, ...)
}
\arguments{
  \item{browserName}{Name of the browser to use for
  testing. See ?RSelenium::remoteDriver for details.}

  \item{dir}{character string with the path to animint's
  source code. Defaults to current directory}

  \item{port}{port number used for local file server}

  \item{...}{list of additional options passed onto
  RSelenium::remoteDriver}
}
\value{
invisible version of the process ID of a child R session
running a local file server.
}
\description{
Initiates a local file server and remote driver.
}
\seealso{
\link{tests_run()}
}

