% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{tests_run}
\alias{tests_run}
\title{Run animint tests}
\usage{
tests_run(dir = ".", filter = NULL)
}
\arguments{
\item{dir}{character string with the path to animint's source code. Defaults to current directory}

\item{filter}{If not NULL, only tests with file names matching
this regular expression will be executed. Matching will take on the
file name after it has been stripped of "test-" and ".r".}
}
\description{
Convenience function for running animint tests.
}
\examples{

\dontrun{
# run tests in test-rotate.R with Firefox
tests_init("firefox")
tests_run(filter = "rotate")
# clean-up
tests_exit()
}

}

