% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/animint.R
\name{toRGB}
\alias{toRGB}
\title{Convert R colors to RGB hexadecimal color values}
\usage{
toRGB(x)
}
\arguments{
  \item{x}{character}
}
\value{
hexadecimal color value (if is.na(x), return "none" for
compatibility with JavaScript)
}
\description{
Convert R colors to RGB hexadecimal color values
}

