\name{UStornadoes}
\alias{UStornadoes}
\docType{data}
\title{
Tornadoes in the United States from 1950 to 2012
}
\description{
  Each row documents 1 tornado.
}
\usage{data(UStornadoes)}
\format{
  A data frame with 41620 observations on the following 32 variables.
  \describe{
    \item{\code{fips}}{a numeric vector}
    \item{\code{ID}}{a numeric vector}
    \item{\code{year}}{a numeric vector}
    \item{\code{month}}{a numeric vector}
    \item{\code{day}}{a numeric vector}
    \item{\code{date}}{factor}
    \item{\code{time}}{a numeric vector}
    \item{\code{tz}}{a numeric vector}
    \item{\code{state}}{factor}
    \item{\code{state.tnum}}{a numeric vector}
    \item{\code{f}}{a numeric vector}
    \item{\code{injuries}}{a numeric vector}
    \item{\code{fatalities}}{a numeric vector}
    \item{\code{propertyLoss}}{a numeric vector}
    \item{\code{cropLoss}}{a numeric vector}
    \item{\code{startLat}}{a numeric vector}
    \item{\code{startLong}}{a numeric vector}
    \item{\code{endLat}}{a numeric vector}
    \item{\code{endLong}}{a numeric vector}
    \item{\code{trackLength}}{a numeric vector}
    \item{\code{trackWidth}}{a numeric vector}
    \item{\code{numStatesAffected}}{a numeric vector}
    \item{\code{stateNumber}}{a numeric vector}
    \item{\code{segmentNumber}}{a numeric vector}
    \item{\code{FipsCounty1}}{a numeric vector}
    \item{\code{FipsCounty2}}{a numeric vector}
    \item{\code{FipsCounty3}}{a numeric vector}
    \item{\code{FipsCounty4}}{a numeric vector}
    \item{\code{TotalPop2012}}{a numeric vector}
    \item{\code{LandArea}}{a numeric vector}
    \item{\code{TornadoesSqMile}}{a numeric vector}
    \item{\code{weight}}{a numeric vector}
  }
}
\source{
NOAA SVRGIS data (Severe Report Database + Geographic Information System)
http://www.spc.noaa.gov/gis/svrgis/
}
\keyword{datasets}
