% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/animint.R
\name{varied.chunk}
\alias{varied.chunk}
\title{Extract subset for each data.frame in a list of data.frame}
\usage{
varied.chunk(df.list, cols, nest_order)
}
\arguments{
\item{df.list}{list of data.frame.}

\item{cols}{cols that each data.frame would keep.}

\item{nest_order}{columns used for d3 nest.}
}
\value{
list of data.frame.
}
\description{
Extract subset for each data.frame in a list of data.frame
}

