\name{WorldBank}
\alias{WorldBank}
\docType{data}
\title{
  Demographics by country from 1960 to 2012
}
\description{
  Each row is one year of demographics for one country.
}
\usage{data(WorldBank)}
\format{
  A data frame with 11342 observations on the following 15 variables.
  \describe{
    \item{\code{iso2c}}{a character vector}
    \item{\code{country}}{a character vector}
    \item{\code{year}}{a numeric vector}
    \item{\code{fertility.rate}}{a numeric vector}
    \item{\code{life.expectancy}}{a numeric vector}
    \item{\code{population}}{a numeric vector}
    \item{\code{GDP.per.capita.Current.USD}}{a numeric vector}
    \item{\code{15.to.25.yr.female.literacy}}{a numeric vector}
    \item{\code{iso3c}}{factor}
    \item{\code{region}}{factor}
    \item{\code{capital}}{factor}
    \item{\code{longitude}}{factor}
    \item{\code{latitude}}{factor}
    \item{\code{income}}{factor}
    \item{\code{lending}}{factor}
  }
}
\source{
Copied from the googleVis package.
}
\keyword{datasets}
