% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testFunctions.R
\name{animint2HTML}
\alias{animint2HTML}
\title{Apply `animint2dir` to a list ggplots and extract the (rendered) page source via RSelenium}
\usage{
animint2HTML(plotList)
}
\arguments{
  \item{plotList}{A named list of ggplot2 objects}
}
\description{
Apply `animint2dir` to a list ggplots and extract the
(rendered) page source via RSelenium
}

