% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/animint.R
\name{getLegendList}
\alias{getLegendList}
\title{Function to get legend information from ggplot}
\usage{
getLegendList(plistextra)
}
\arguments{
\item{plistextra}{output from ggplot2::ggplot_build(p)}
}
\value{
list containing information for each legend
}
\description{
Function to get legend information from ggplot
}

