\name{gg2animint}
\alias{gg2animint}
\title{gg2animint}
\description{Convert a list of ggplots to an interactive animation.}
\usage{gg2animint(plot.list, out.dir = tempfile())}
\arguments{
  \item{plot.list}{List of ggplots with showSelected, time, and/or clickSelects aes.}
  \item{out.dir}{Directory to store html/js/csv files.}
}

\value{Nothing.}

\author{Toby Dylan Hocking}





