% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/animint.R
\name{gg2animint}
\alias{gg2animint}
\title{animint2dir}
\usage{
gg2animint(...)
}
\arguments{
\item{...}{passed to animint2dir}
}
\value{
same as animint2dir
}
\description{
Deprecated alias for animint2dir.
}
\author{
Toby Dylan Hocking
}

