\name{gg2list}
\alias{gg2list}
\title{Convert a ggplot to a list. Called from gg2animint().}
\usage{
gg2list(p)
}
\arguments{
  \item{p}{ggplot2 plot}
}
\value{
list representing a ggplot, with elements, ranges, axes,
scales, geoms, and options
}
\description{
Convert a ggplot to a list. Called from gg2animint().
}
\seealso{
\code{\link{gg2animint}}
}

