% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/animint.R
\name{is.clickSelects}
\alias{is.clickSelects}
\title{Test if aesthetics are clickSelects.}
\usage{
is.clickSelects(x)
}
\arguments{
  \item{x}{character vector.}
}
\value{
logical vector
}
\description{
Test if aesthetics are clickSelects.
}
\author{
Susan VanderPlas
}

