% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{knit_print.animint}
\alias{knit_print.animint}
\title{Insert an interactive animation into an R markdown document using a customized print method.}
\usage{
knit_print.animint(x, options, ...)
}
\arguments{
\item{x}{named list of ggplots and option lists to pass to \code{animint2dir}.}

\item{options}{knitr options.}

\item{...}{placeholder.}
}
\description{
Insert an interactive animation into an R markdown document using a customized print method.
}
\author{
Carson Sievert
}
\references{
https://github.com/yihui/knitr/blob/master/vignettes/knit_print.Rmd
}

