\name{parsePlot}
\alias{parsePlot}
\title{Convert a ggplot to a list. Called from gg2animint().}
\usage{
parsePlot(meta)
}
\arguments{
  \item{meta}{environment with previously calculated plot
  data, and a new plot to parse, already stored in plot and
  plot.name.}
}
\value{
nothing, info is stored in meta.
}
\description{
Convert a ggplot to a list. Called from gg2animint().
}
\seealso{
\code{\link{gg2animint}}
}

