% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parsePlot}
\alias{parsePlot}
\title{Convert a ggplot to a list.}
\usage{
parsePlot(meta)
}
\arguments{
  \item{meta}{environment with previously calculated plot
  data, and a new plot to parse, already stored in plot and
  plot.name.}
}
\value{
nothing, info is stored in meta.
}
\description{
Convert a ggplot to a list.
}

