% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitr.R
\name{renderAnimint}
\alias{renderAnimint}
\title{Create an animint output element}
\usage{
renderAnimint(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
  \item{expr}{An expression that creates a list of ggplot
  objects.} \item{env}{The environment in which to evaluate
  \code{expr}.} \item{quoted}{Is expr a quoted expression
  (with \code{quote()})? This is useful if you want to save
  an expression in a variable.}
}
\description{
Shiny server output function customized for animint plots
(similar to \code{shiny::plotOutput} and friends).
}
\seealso{
http://shiny.rstudio.com/articles/building-outputs.html
}

