% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animint.R
\name{saveChunks}
\alias{saveChunks}
\title{Split data set into chunks and save them to separate files.}
\usage{
saveChunks(x, meta)
}
\arguments{
\item{x}{data.frame.}

\item{meta}{environment.}
}
\value{
recursive list of chunk file names.
}
\description{
Split data set into chunks and save them to separate files.
}
\author{
Toby Dylan Hocking
}

