% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/animint.R
\name{saveCommonChunk}
\alias{saveCommonChunk}
\title{Save the common columns for each tsv to one chunk}
\usage{
saveCommonChunk(x, vars, nest_order, meta)
}
\arguments{
\item{x}{data.frame.}

\item{vars}{character vector of variable names to split on.}

\item{nest_order}{character vector of nest variable names.}

\item{meta}{environment.}
}
\value{
a list of data.frames comprised of varied columns for each tsv.
}
\description{
Save the common columns for each tsv to one chunk
}

