% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{tests_exit}
\alias{tests_exit}
\title{Kill child process(es) that may have been initiated in animint testing}
\usage{
tests_exit()
}
\description{
Read process IDs from a file and kill those process(es)
}
\seealso{
\link{tests_run}
}

