% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testHelpers.R
\name{tests_run}
\alias{tests_run}
\title{Run animint tests}
\usage{
tests_run(envir, filter = NULL)
}
\arguments{
  \item{envir}{an environmnent.}

  \item{filter}{If not NULL, only tests with file names
  matching this regular expression will be executed.
  Matching will take on the file name after it has been
  stripped of "test-" and ".r".}
}
\description{
Convenience function for running animint tests.
}
\examples{
\dontrun{
# run tests in test-rotate.R with Firefox
env <- tests_init("firefox")
tests_run(env, filter = "rotate")
tests_exit(env)
}
}

