% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animint.R
\name{toRGB}
\alias{toRGB}
\title{Convert R colors to RGB hexadecimal color values}
\usage{
toRGB(x)
}
\arguments{
\item{x}{character}
}
\value{
hexadecimal color value or "transparent" if is.na
}
\description{
Convert R colors to RGB hexadecimal color values
}

