% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animint.R
\name{varied.chunk}
\alias{varied.chunk}
\title{Extract subset for each data.frame in a list of data.frame}
\usage{
varied.chunk(df.or.list, cols)
}
\arguments{
\item{df.or.list}{a data.frame or a list of data.frame.}

\item{cols}{cols that each data.frame would keep.}
}
\value{
list of data.frame.
}
\description{
Extract subset for each data.frame in a list of data.frame
}

